+++ lib/coinbase/staking_operation.rb
@@ -89,7 +89,9 @@ module Coinbase
     # Returns a new StakingOperation object.
     # @param model [Coinbase::Client::StakingOperation] The underlying StakingOperation object
     def initialize(model)
-      from_model(model)
+      @model = model
+      @transactions ||= []
+      update_transactions(model.transactions)
     end
 
     # Returns the Staking Operation ID.
@@ -116,18 +118,48 @@ module Coinbase
       @model.status
     end
 
+    # Returns whether the Staking Operation is in a terminal state.
+    # @return [Boolean] Whether the Staking Operation is in a terminal state
+    def terminal_state?
+      failed? || completed?
+    end
+
+    # Returns whether the Staking Operation is in a failed state.
+    # @return [Boolean] Whether the Staking Operation is in a failed state
+    def failed?
+      status == 'failed'
+    end
+
+    # Returns whether the Staking Operation is in a complete state.
+    # @return [Boolean] Whether the Staking Operation is in a complete state
+    def completed?
+      status == 'complete'
+    end
+
+    # Returns a String representation of the Staking Operation.
+    # @return [String] a String representation of the Staking Operation
+    def to_s
+      Coinbase.pretty_print_object(
+        self.class,
+        id: id,
+        status: status,
+        network_id: network.id,
+        address_id: address_id
+      )
+    end
+
     # Returns the Wallet ID of the Staking Operation.
     # @return [String] The Wallet ID
     def wallet_id
       @model.wallet_id
     end
 
-    # Waits until the Staking Operation is completed or failed by polling its status at the given interval. Raises a
-    # Timeout::Error if the Staking Operation takes longer than the given timeout.
+    # Waits until the Staking Operation is completed or failed by polling its status at the given interval.
     # @param interval_seconds [Integer] The interval at which to poll, in seconds
     # @param timeout_seconds [Integer] The maximum amount of time
     #   to wait for the StakingOperation to complete, in seconds
     # @return [StakingOperation] The completed StakingOperation object
+    # @raise [Timeout::Error] if the Staking Operation takes longer than the given timeout.
     def wait!(interval_seconds = 5, timeout_seconds = 3600)
       start_time = Time.now
 
@@ -135,11 +167,49 @@ module Coinbase
         reload
 
         # Wait for the Staking Operation to be in a terminal state.
-        break if status == 'complete'
+        break if terminal_state?
 
         raise Timeout::Error, 'Staking Operation timed out' if Time.now - start_time > timeout_seconds
 
         self.sleep interval_seconds
+      end
+
+      self
+    end
+
+    # Complete helps the Staking Operation reach complete state, by polling its status at the given interval, signing
+    # and broadcasting any available transaction.
+    # @param key [Eth::Key] The key to sign the Staking Operation transactions with
+    # @param interval_seconds [Integer] The interval at which to poll, in seconds
+    # @param timeout_seconds [Integer] The maximum amount of time
+    #   to wait for the StakingOperation to complete, in seconds
+    # @return [StakingOperation] The completed StakingOperation object
+    # @raise [Timeout::Error] if the Staking Operation takes longer than the given timeout.
+    def complete(key, interval_seconds: 5, timeout_seconds: 600)
+      start_time = Time.now
+
+      loop do
+        @transactions.each_with_index do |transaction, i|
+          next if transaction.signed?
+
+          transaction.sign(key)
+          @model = Coinbase.call_api do
+            stake_api.broadcast_staking_operation(
+              wallet_id,
+              address_id,
+              id,
+              { signed_payload: transaction.raw.hex, transaction_index: i }
+            )
+          end
+        end
+
+        return self if terminal_state?
+
+        reload
+
+        raise Timeout::Error, 'Staking Operation timed out' if Time.now - start_time > timeout_seconds
+
+        sleep interval_seconds
       end
 
       self
@@ -184,7 +254,9 @@ module Coinbase
         end
       end
 
-      from_model(@model)
+      update_transactions(@model.transactions)
+
+      self
     end
 
     # Fetches the presigned_voluntary exit messages for the staking operation
@@ -231,14 +303,22 @@ module Coinbase
       @wallet_stake_api ||= Coinbase::Client::WalletStakeApi.new(Coinbase.configuration.api_client)
     end
 
-    def from_model(model)
-      @model = model
-      @status = model.status
-      @transactions = model.transactions.map do |transaction_model|
-        Transaction.new(transaction_model)
+    def update_transactions(transactions)
+      # Only overwrite the transactions if the response is populated.
+      return unless transactions && !transactions.empty?
+
+      # Create a set of existing unsigned payloads to avoid duplicates.
+      existing_unsigned_payloads = Set.new
+      @transactions.each do |transaction|
+        existing_unsigned_payloads.add(transaction.unsigned_payload)
       end
 
-      self
+      # Add transactions that are not already in the transactions array.
+      transactions.each do |transaction_model|
+        unless existing_unsigned_payloads.include?(transaction_model.unsigned_payload)
+          @transactions << Transaction.new(transaction_model)
+        end
+      end
     end
   end
 end